requirements.txt
black
pyyaml~=6.0
jinja2~=3.1.2
fastapi~=0.95.0
giteapy~=1.0.8
uvicorn
python-dotenv~=1.0.0
pydantic~=1.10.6
